Process keys not in unicodemap using keycode table
Previously, if a key press had an associated Unicode value and that
value was not found in the unicodemap, the press would be ignored.
This is not what we want for ColecoVision, where we selected some of
the keys (like WASD) based on their position on the keyboard rather
than their Unicode value.