Initialize MSXMusicWX::control in constructor
First I (wrongly) thought this was done via the reset() call in the base class,
but virtual function calls in the base constructor can't call methods in the
subclass.

In practice the initialization is done correctly anyway because the default
implementation of powerOn() calls reset(), and powerOn() is called before the
machine actually starts running. But it's cleaner to not rely on that.

https://scan7.coverity.com/reports.htm#v13443/p10370/fileInstanceId=45517458&defectInstanceId=9573575&mergedDefectId=1463899