Added unittest for sha1 utility
Thanks to Maarten for extending the build infrastructure.

This patch replaces the dummy sha1 test with a real sha1 unittest.

To build the tests do:
  > make OPENMSX_FLAVOUR=unittest -j<n>
To run just execute the generated executable
  > openmsx
Or pass '--help' to see the possible options.

Since this is the first (non-dummy) unittest, I'd like to point to the catch2
tutorial:
  https://github.com/catchorg/Catch2/blob/master/docs/tutorial.md