Allow for software controllable sound-chip volume
(Will be) used in Carnivore2.

The Carnivore2 device includes (in FPGA) a SCC and FM-PAC sound chip. It also
has a register that allows to change the total volume of those sound chips (I
mean not the regular per-channel SCC/YM2413 volume registers, but an additional
per-chip one). Thus the volume is controllable via MSX software.

Before this patch openMSX had no mechanism yet to allow the MSX software to
change the volume of a sound chip. (It was already possible via various
settings or via the configuration file).