Implemented 'global read' mechanism
Similar to the already existing 'global write' stuff. This will be used in
emulation of the 'Carnivore2' device.

To detect when the MSX machine is reset (soft-reset) 'Carnivore2' monitors the
MSX bus for reads in any slot (normally a device should check both the
address(-range) and the slot-select-signal). It does this to:
 - detect execute (= read + active M1-signal) from address 0x0000
 - detect read from address 0x4000