Fixed exception safety problem in Carnivore2
(Before this patch) during the Carnivore2 constructor we call
MSXMapperIO::registerMapper(). Such a call must always be paired with a
corresponding call to MSXMapperIO::unregisterMapper(). This 2nd call is located
in the Carnivore2 destructor. So far so good. Though when an exception occurs
after the registerMapper() call but before the Carnivore2 constructor is fully
finished, there will be no call to unregisterMapper().

Fixed by encapsulating the register/unregister call-pair into a new
MSXMapperIOClient helper class. And used this helper in both Carnivore2 and
MSXMemoryMapper.

Bonus: the new code doesn't store a reference the to MSXMapperIO object, making
the size of a Carnivore2 and a MSXMemoryMapper object a bit smaller.