Fix (workaround) replay problem on Coleco machines
To reproduce the problem:
 - Start (any) game in the 'ColecoVision' machine.
 - In the console execute 'reverse goto 0'.
This works as expected when openMSX was started from the command line. But when
openMSX was started via catapult, the game doesn't start anymore. It's as if
the game roms isn't inserted anymore after the reverse. Some debugging showed
this is because the wrong slots are selected.

The difference between these two scenario's is how the machine gets
powerUp()'ed.
* command line:
  - Machine gets instantiated, and soon after powerUp() gets called.
* catapult:
  - Machine gets instantiated, and only later powerUp() gets called (via a 'set
    power on' command).
This influences when the initial 'reverse' snapshot gets made. In the first
scenario powerUp() is already called, in the second scenario it is not.

But how is powerUp() connected to wrong slots being selected? It turns out the
MSXCPUInterface constructor doesn't fully initialize the object. More
specifically the slot-state is not yet initialized in the constructor. Instead
we rely on the reset() method to finish the initialization. So in the
cmdline-scenario the initial snapshot has a fully initialized slot-status, in
the catapult-scenario it has not. MSXMotherBoard::powerUp() calls
MSXCPUInterface::reset(), so everything should still be fine ... except that
power on (or power off) events are not included in the replay.

For now I only applied a workaround that fully initializes MSXCPUInterface in
the constructor, so it doesn't matter that powerUp() doesn't get called. Or at
least it doesn't matter anymore for the MSXCPUInterface class, but there might
be other classes where is does still matter. So the real fix should be to
properly handle power on/off in replays.