Replace 'unplug' event with 'plug-empty'
Consider the following scenario:
 - plug joyporta joystick1
 - unplug joyporta
 - plug joyporta joystick1

Before this patch that would send these events:
  <update type="plug" name="joyporta">joystick1</update>
  <update type="unplug" name="joyporta"></update>
  <update type="plug" name="joyporta">joystick1</update>

Except that the 3rd event wouldn't be send because the data is identical to the
previously send 'plug' event. And we filter out repeated events with the same
data.

This problem arises because we use two different events to signal changes in a
single state (what's plugged in a certain connector). After realizing this, the
fix easy.

After this patch we send the following events:
  <update type="plug" name="joyporta">joystick1</update>
  <update type="plug" name="joyporta"></update>
  <update type="plug" name="joyporta">joystick1</update>
So we replace the 'unplug' event with a 'plug' event that has an empty string
as the 'pluggable' name.

Initially we considered still sending the 'unplug' event (in addition to the
new empty-plug event) for backwards compatibility. But it turned out that
catapult (both the wxWidget and the python version) and the openmsx debugger
are already working fine with this new schema.

This patch also changes the name of the dummy pluggable from "--empty--" to the
empty string. (The dummy pluggable is plugged-in when nothing is physically
plugged in the connector. This simplifies certain parts of the implementation).

One consequence of this is that the 'plug' command now shows an empty string
instead of '--empty--' for connectors that have nothing plugged in. That's also
a backwards-incompatible change, and we did need to adapt a few Tcl scripts for
this.