Fix compilation in debug mode
The intrinsic function _mm_slli_si128() needs a constant expression for its 2nd
parameter. For some reason gcc-7.3 doesn't see the variable
   static const int TMP = sizeof(__m128i) - BYTES;
as a constant expression. Not even when written like this
   static constexpr int TMP = sizeof(__m128i) - BYTES;
At least in debug mode, in a normal build it works fine.

Worked around it by making the expression a (defaulted) template parameter.