Fix crash when formatting an empty (filesize=0) disk image
Reported by Patrick. Steps to reproduce:
 - Create an empty file test.dsk.
 - Use that disk image and do a "CALL FORMAT" in BASIC.
 - Crash.

The problem was more general, it could (possibly?) also occur on other write
errors. But should be fixed now. (The write error was already handled (passed
to the MSX), but something went wrong in the invalidation of the cached track
info in the disk drive emulation).

BTW: zero-sized disk images are not very useful, so you could say we shouldn't
allow them. Though they can be useful in testing to simulate a disk that gives
read/write errors everywhere.