Added strCat() and strAppend() utility functions
Inspired by google's absl::StrCat() utilities. See
   https://abseil.io/blog/20171023-cppcon-strcat
   https://github.com/abseil/abseil-cpp/blob/master/absl/strings/str_cat.h

In openMSX strCat() can replace the StringOp::Builder() utility that we had.
strCat() has a nicer API compared to StringOp::Builder(). Compare:
   error(strCat("Error: expected ", e, " but got ", r, '.'));
vs
   error(StringOp::Builder() << "Error: expected " << e << " but got " << r << '.');
or even
   error("Error: expected " + std::to_string(e) + " but got " + std::to_string(r) + '.');

strCat() is also faster than StringOp::Builder() because:
- strCat():
  * First pass figures out the string-length of each chunk (and sums those).
  * We allocate a buffer with the exact required size.
  * Second pass actually converts each chunk to string.
- StringOp::Builder():
  * Only makes a single pass. It doesn't know the final size upfront, therefor
    it might have to reallocate (possibly more than once).
It would be possible to use the same 2-pass approach in StringOp::Builder(),
but it's harder to implement there because not all arguments are visible in a
single function call.

There's also a strAppend() function that appends stuff to an existing string.
For example:
   strAppend(s, "Another thing: ", thing, '\n');
instead of:
   s += strCat("Another thing: ", thing, '\n');
or even:
   s += "Another thing: " + to_string(thing) + '\n';
strAppend() avoids the overhead of first building a temporary and only then
appending that to the result. Instead it directly builds the result in the
final location.

This patch only adds the new utility (+unittest) but doesn't use it yet. It
also doesn't remove the old functionality yet. So temporarily we have some code
duplication between the two. E.g. the helper functions for fast int->string
conversion are now copy/pasted between the two implementations.