Removed (mixed) string/string_ref operator+ functions
Our string_ref class is modeled after the (c++17) std::string_view class. The
intention is to (in the future) replace our version with the standard version.
But we have some functions that are not part of the standard. Specifically we
have:
   std::string operator+(string_ref x, string_ref y);
   std::string operator+(char x, string_ref y);
   std::string operator+(string_ref x, char y);

Now that we added strCat() we have an easy replacement for these 3 functions.
An earlier patch already rewrote all uses of these functions to strCat(). This
patch actually removes the functions.