Renamed string_ref to string_view
'string_ref' started as a proposal for a future c++ version. In c++17 it has
become part of standard c++. But along the way it got renamed to 'string_view'.
(And I agree this is a better name).

In openMSX we started from an early version of this proposal (which was still
called 'string_ref). This patch also renames our version to string_view.

The intention is that in the future we can drop our version and instead use
std::string_view (so best stick close to it). But before we can do that we
first need to be able to count on a c++17 compiler to compile openMSX (we've
only just switched to c++14).