Move omission of ALSA on non-Linux systems from Make to Python
The idea is to have the build configuration in Python and the execution
in Make.

Note that the platform check switched from using the value of TRIPLE_OS
to OPENMSX_TARGET_OS. This has no effect right now, since the only
platform for which these two are different is OpenDingux, which has
ALSA as a system library. But it will be relevant for the updated
Android support.