Improve probe for libraries that try to intercept main()
Usually when main() is intercepted, it is provided by the library.
However, this is not the case for Android, where the entry point
for an application is in Java.

Previously I tried linking the test executable as a shared library.
While that does remove the errors, it also defeats the purpose of
the probe, as missing symbols are no longer errors.