Build appdata.zip from bindist/install/share content
The content of this dir is explicitly put there for distribution, so it's only
logical to use it also for the Android distribution.
Also removed the check whether it is still valid. Rebuilding it is a matter of
seconds, so just always create it.
Lastly, do not copy C-BIOS, as it's already there.