integrate OPL4 improvements made by Valley Bell
Thanks a lot Valley Bell!

I reviewed all changes, and they all look fine. Though I cannot be 100% sure
there won't be any regressions. So please test if you can.

One thing that did change (unintentionally according to Valley Bell) is the
content of the am_depth[] array. Though it's not clear that this is good or bad
(the corresponding change in the dl_tab[] array _is_ intended. We've fixed it
as good as we can for now, but needs more investigation.