Implement YMF262 mix levels
That is OPL4 register 0xF8. This controls the attenuation factors for the
left/right audio channel for the whole FM output. (Register 0xF9 was already
implemented, mix levels for the wave part).

Internally implemented via the 'setSoftwareVolume()' mechanism (but now
extended to allow a different factor for left and right). This saves 2
multiplications per sample compared to implementing it in the
generateChannels() method (as was done before for the wave part).

I verified that before/after this patch, with the default settings for
registers 0xf8/0xf9 the output volume remained the same. (I'm not 100% sure
the balance between wave and FM is the same as on the real chip).