fix Wformat-truncation warnings in gcc-8
At two places in the code we print a formatted date
  YYYY-MM-DD
As long as the year is less than 9999 this will always fit in 10+1 characters
(+1 for zero-terminator). But of course the compiler doesn't understand
calendars and warns that the output might get truncated (e.g. because the year
is -123456 or because it doesn't know a month is at most 2 digits).

Theoretically the compiler is right, but in practice there's no problem. But
the fix is easy enough: just enlarge the buffer (and has no performance impact
at all: it's a local buffer on the stack).

After this patch, a debug/devel build on linux with gcc-8 is warning-free again.