fix bug #1117: Reverse issue with GFX9000
The V9990CmdEngine class has a member variable 'cmdMode'. In a way this is a
cached value: it can always be recalculated from the V9990 display mode.
Therefor in a savestate the value of this variable is not stored but instead
it's recalculated on load. But this is of course only correct if the display
mode is already correctly restored _before_ we perform this recalculation.