Remove pre-c++14 constexpr workaround(s)
Replace CONSTEXPR macro with constexpr keyword. (CONSTEXPR was #defined as
'constexpr' in c++14 mode or as the empty string in pre-c++14 mode).

Remove workaround for gcc-5 constexpr bug. This means the minimum required gcc
version is now gcc-6. I updated the manual for this.

Simplified CRC16 stuff by using constexpr:
- Replace hardcoded lookup table by a version that gets calculated at compile
  time.
- Change CRC16::init() implementation from template-meta-programming to a
  'normal' (constexpr) function. This is much simpler, and I verified that the
  calculation for the initializing value of a CRC16 object still fully happens
  at compile time.