Replace 'array_ref' utility with 'span'
Our 'array_ref' utility was modeled after a proposal for c++ standardization.
That proposal has evolved quite a bit, and there's a good chance it will be
accepted in the c++20 standard. The most important changes compared to earlier
versions of the proposal are:

- It was renamed from 'array_ref' to 'span'.
- Both dynamic- and fixed-size spans are now possible. (Remember under the hood
  array_ref was basically a pointer and a size.) For a fixed-size span it's not
  necessary to explicitly store the size, instead it's part of the type itself.
- Better support for const vs non-const elements.
  E.g. span<int> vs. span<const int>.

This patch replaces our 'array_ref' utility with (still our own
re-implementation of) a 'span' utility. Apart from that, the majority of this
patch are mechanical renames (array_ref->span). In the future we can drop our
implementation and replace 'span' with 'std::span'.