fix hang on invalid <initialContent> tag
Problem reported by Manuel:
- Inside the <VDP> tag in any machine config add:
     <initialContent encoding="hex">FG</initialContent>
  (intentional invalid HEX value).
- Using this machine hangs openMSX.

The current HexDump::decode() function ignores non-hex characters (all other
than 0-9,a-f,A-F), this was sort of intentional to allow for spaces and/or
newlines between the digits. But maybe we should do more strict checking?

Anyway, the reason for the hang was that the hex pattern has zero length (no
valid hex bytes). And filling all VRAM by repeating that pattern takes a very
long time.

Fixed by checking for a zero-length pattern. But as explained above, maybe we
should also do more checks on the pattern itself? I left that part for a later
patch.