Fixed bug#497 Ashiguine 3 sprite bug
Sprites with CC=1 that are not preceded by a sprite with CC=0 are not visible,
but they DO contribute towards the limit of maximum 8 sprites per line.

The old spriteChecker (wrongly) skipped sprites with CC=1 if there wasn't any
prior sprite (with CC=0) in the spriteBuffer[] array yet. This simplifies the
later sprite rendering code, but as a side effect this also (unintentionally)
changes the logic to detect more than 8 sprites per line. Fixed by moving the
'invisible CC=1 sprite'-check to the sprite rendering side.

--

More details about the 'Ashiguine 3' bug:

a) It went wrong in openMSX on most machines.
b) It works fine in openMSX on CBIOS machine.
c) It works fine in blueMSX.

I verified that:
- In all 3 cases the VDP registers have the same values.
- Between case a) and b) the content of the VRAM is different. This difference
  comes from the standard MSX bootlogo versus the CBIOS bootlogo.
- Cases a) and c) have the same VRAM content.
So in cases a) and c) the VDP+VRAM is in exactly the same state, therefor
openMSX must be doing something wrong related to sprite rendering.

The bug triggers in screen mode 5, 16x16 sprites. The sprite tables are:
- sprite pattern table:   [0x7000, 0x77ff]
- sprite color table:     [0x7980, 0x7b7f]
- sprite attribute table: [0x7b80, 0x7bff]
The VRAM region [0x7000, 0x8000) is filled with the following pattern:
  [0x..00, 0x..ef] = 0x55  left-over from the bootlogo (=0x00 for CBIOS)
  [0x..f0, 0x..ff] = 0x00  set to 0 by 'Ashiguine 3'
If you decode this you get:
 sprites  0- 6  y=85, CC=1, color=5
 sprite     7   y=85, CC=0, color=0
 sprites  8-14  y=85, CC=1, color=5
 sprites   15   y=85, CC=0, color=0
 ...

Before this patch openMSX would (on lines 85-100):
- Skip sprites 0-6 because they have CC=1 and there wasn't any sprite yet with
  CC=0.
- Sprite 7 has CC=0, so it's not skipped, but because it has color=0 it's drawn
  fully transparent.
- Sprites 8-14 have CC=1, they are not skipped because we've already seen
  sprite 7. These sprites show up as the wrong garbage sprites.
- At this point we have 8 visible sprites.

After the fix openMSX does (on lines 85-100):
- Pick sprites 0-6, but because they have CC=1 they aren't drawn.
- Pick sprite 7, but because it has color=0 it's fully transparent.
- At this point we have 8 visible sprites, so sprites 8-14 aren't drawn
  anymore.