Fixed possible crash in hq(lite) scalers in combination with interlaced screen modes
Report by PatrickR here:
  https://www.msx.org/forum/msx-talk/openmsx/please-help-testing-upcoming-openmsx-release?page=4

These scalers execute some 'edge-calculation' (on the CPU) and upload that
result as a texture to the GPU. Though in case of interlace the height of that
'edge-buffer' wasn't large enough.

Fixed by disabling the edge calculation for interlaced screen modes. It's not
useful to increase the height of the edge buffer because the scaler anyway only
works for input screens with resoltion 320x240 (256x212 msx screen mode). This
actually also means the bug only triggered for interlaced msx screens with a
x-resolution of 256 (instead of 512) pixels (this is relatively rare).