small cleanups
Found when enabling clang -Weverything and sifting through the long list of
mostly useless warnings. In other words, except for the tiny changes made in
this commit, -Weverything didn't find anything useful. So there are no useful
extra warnings we can enable on clang. (There's a reason why these warnings are
not part of -Wall).