[2/2] Move AY8910::Generator output from base to subclasses
For the 'ToneGenerator' (sub)class this means move the member variable from the
base class to this (sub)class and changing the type from 'unsigned' to 'bool'
(because it can only have value 0 or 1, verified in the previous patch).

For the 'NoiseGenerator' (sub)class this means removing it from the base class
and don't add anything to this (sub)class. Instead the value can easily be
calculated from the expression 'random & 1'. I verified that the generated code
for the inner loop in generateChannels() is now a few instructions shorter (the
variants that have noise enabled).