Use the correct LFSR formula for AY8910
The original AY8910 emulation in openMSX was based on mame-0.37b (2001). That
uses the following formula (a LFSR) for the noise generator:
     if (random & 1) random ^= 0x28000;
     random >>= 1;

Current versions of mame use a different formula:
     if (random & 1) random ^= 0x24000;
     random >>= 1;
(or the equivalent in Fibonacci- instead of Galois-form)

I verified that mame-0.121 (2007) already has the new formula, but
unfortunately the mame project on github doesn't have any older history than
that. So I couldn't verify why this was changed.

Current mame has a comment that the LFSR formula was verified on real AY-3-8910
and YM2149 chips. Additionally the new formula is 'maximal-length' while the old
one was not (and this always bothered me). So from an engineering point of view
there are compelling reasons to prefer the new over the old formula.

I asked Eugeny Brychkov to make some measurements of a real PSG while running
the following BASIC program (context: he's working on adding PSG emulation in
GR8NET, and while debugging his PSG implementation he was already comparing his
emulated PSG with the real PSG).
   10 SOUND 6,31
   20 SOUND 7,183
   30 SOUND 8,15
   (this makes the noise output easily visible)
Part of the measurement shows this noise sequence:
  0,1,1,0,1,0,0,1,0,1,0,0,1,0,1,0,0,0,0,1,0,0,0,1,1,0,0,0,1,1,0,1,0,1,0,0,1,1,1
And when I run that sequence through the Berlekampi-Massey algorithm
  https://en.wikipedia.org/wiki/Berlekamp%E2%80%93Massey_algorithm
I indeed get the LFSR that corresponds to the 'new' formula.

Anyway, it's good that we fixed the formula, but in practice this won't make
much difference: random noise still sounds like random noise. I checked: I
really couldn't hear any difference before/after this patch e.g. when running
the above BASIC program.