Made memory mapper ports readable on MegaFlashRomSCCPlusSD
I got an email from Christian Dorazio asking why, in openMSX, reading from the
memory mapper registers (0xfc-0xff) always returns the value from the internal
memory mapper rather than from the MegaFlashRomSCCPlusSD device. And whether on
a real MSX the behavior is the same.

The answer was that MegaFlashRomSCCPlusSD does not implement reading from these
registers at all. As documented by Manuel Pazos (creator of the cartridge):
  Since mapper ports must not be read, as stated on MSX Technical Handbook, and
  mapper ports as read only, as stated on MSX Datapack, all read operations on
  these ports will not return any value.

However from private emails between Christian and Manual it seems that initially
MegaFlashRomSCCPlusSD indeed did not implement reading. But upon user request,
this feature was added later.

Thus this patch adds reading the memory mapper registers to the
MegaFlashRomSCCPlusSD device. I did not bother making this behavior
configurable.

Implementation detail:
To solve an order-of-initialization problem, I had to change the interface of
the MSXMapperIOClient helper class slightly. Before it got the MSXMotherBoard
via some CRTP-trick, now it receives it as a constructor parameter and stores
it as a member variable.