reintroduce the unicode field on key-events
Added an heuristic to transform the separate SDL2 text and key events back into
SDL1-style combined SDL key event with unicode field.

See comments in the code for a more detailed description of this heuristic.

This was needed for the openMSX keyboard emulation where we try to remap host
keys onto the corresponding msx key presses that produce the same character.
E.g. typing '(' on the host keyboard with AZERTY keyboard layout (press the '5'
key without holding shift) will result in pressing 'shift'+'9' on a Philips MSX
machine with QWERTY keyboard layout.