Fix 'openmsx_info software <sha1sum>'
There was a bug that caused the 'software' info topic to not be available all
the time. More specifically, before this patch, the topic was created by the
RomDatabase class. But at some point we changed the code to only create the
RomDatabase on demand. And thus the info topic was unintentionally only
available anymore after the RomDatabase was used.

Fixed by unconditionally instantiating the 'SoftwareInfoTopic' object. To do
this the code was moved from RomDatabase.cc to Reactor.cc.