split SDL-text event into multiple openMSX-keydown events
In the early SDL2 port we introduced a TextEvent class that correspond to the
(new in SDL2) TEXTINPUT event. The previous patch already merged part of the
information contained in the TextEvent into the corresponding KeyDown event.
This patch completely eliminates TextEvent, and fully decomposes it into
(possibly multiple) KeyDown events.

The reason for this is that internally in openMSX we have a priority system
when handling keyboard events. That is the openMSX console can block events for
the hotkey subsystem, which in turn can block events for the emulated msx
keyboard. This priority system is much simpler when it only needs to deal with
one type of event. And especially when the same host key press could result in
two generated events (one keydown and one text event).