[6/6] apply fixes suggested by clang-tidy clang-analyzer-optin.cplusplus.VirtualCall
Calling virtual methods from the constructor/destructor of a baseclass can
behave differently than what some people may expect (e.g. if you're more
familiar with other languages than C++): it does not neccesarily invoke the
most derived implementation.

The current code already worked as intended. But because this can be tricky
it's better to be very explict and annotate these calls with the class name.