Clean up CAPS lock handling code
Old SDL 1.2 versions would report the CAPS lock state rather than key
presses and we had code to work around that. Later SDL 1.2 versions
supported an environment variable to request that CAPS be treated like
any other key and we used that to get more reliable behavior.

In SDL2, there is only one behavior: key presses are reported. So we
don't need the code anymore that handles the older behavior.