Restore freetype-config override in 3rdparty
I thought it was no longer needed now that we use pkg-config, but it
turns out the current release of SDL2_ttf only uses pkg-config to find
SDL2 and still requires the freetype-config script to find FreeType.

On my PC, the native freetype-config is found instead and this happens
to not break the build, but it did hide this problem. On machines
without native freetype-config, the build breaks though.