Unset all build tools when building pkg-config
Unlike the library builds, pkg-config is a native tool, so we
shouldn't pass cross tools.

On my PC just unsetting the compiler was enough, but that didn't
work for Manuel. And it's cleaner to unset every anyway.

I guess it would be even better to not export any variables and
instead pass the cross tools explicitly to the cross configures
and Make invocations. Something to consider for the future.