Fixed crash on V9990 overscan modes in combination with the Deflicker filter
The Deflicker filter code has an SSE optimized version (I verified: current
compilers are not yet able to automatically vectorize this code, and the
manually vectorized code is ~4x faster than the plain C++ version).

One limitation of that SSE2 code was that it could only handle line-widths that
are a multiple of the width of a SSE register (=128 bits). Thus for 32bpp it
could only handle widths that are a multiple of 4 pixels.

Usually this limitation is fine ... except for the V9990 overscan modes.

Fixed by processing the bulk of the line with SSE2, and the remaining few
pixels with the fallback C++ code.

There was a similar limitation that SSE code needs 16-byte aligned buffers, but
when using the V9990 overscan modes this alignment is not always guaranteed.

Fixed by using unaligned SSE load/store instructions. On older x86
architectures there was a penalty for using these instructions (compared to
the aligned variants). But on newer architectures the penalty is small or even
zero (at least when the load/store does not cross a cacheline boundary).