[1/15] Added range algorithms
These range-based algorithms (only) delegate to the corresponding iterator-based
version of the same algorithm. These will become part of c++20.

This allows to simplify code like
    std::sort(begin(myVector), end(myVector));
to
    ranges::sort(myVector);

Range algorithms are easier to compose. Later in this series we'll introduce
'views' and that will e.g. allow to rewrite stuff like
    bool anyFooValue(std::map<int, int>& myMap) {
        for (auto& pair : myMap) {
            if (isFoo(pair.second) {
                return true
            }
        }
        return false;
    }
as
    bool anyFooValue(std::map<int, int>& myMap) {
        return ranges::any_of(view::values(myMap),
                              [](auto& value) { return isFoo(value); });
    }