[4/15] Added to_vector() function
This function converts the given range (any kind of range) into a vector. The
type of the elements in the vector is automatically deduced, but it's also
possible to explicitly specify it (e.g. to force a type conversion).

For now this function is of limited use. It will become more useful later in
this patch series (e.g. after view::transform() is introduced).