[12/15] Use view::transform()
Most patches in this series simplify the code by raising the abstraction level.
This patch does the same, but in addition it also improves performance. I'll
explain how that's possible with an example:

version 1:
    std::vector<string_view> OSDTopWidget::getAllWidgetNames() const {
        std::vector<string_view> result;
        for (auto* p : widgetsByName) {
            result.push_back(p->getName());
        }
        return result;
    }

version 2:
    std::vector<string_view> OSDTopWidget::getAllWidgetNames() const {
        return to_vector(view::transform(widgetsByName,
                                         [](auto* p) { return p->getName(); }));
    }

The 1st version starts with an empty vector and then in a loop adds elements
one-by-one. When the capacity of the vector runs out, the internal memory block
will be reallocated. Possibly this happens multiple times when 'widgetsByName'
contains many elements.

The traditional way to improve upon this is to add a line like
    result.reserve(widgetsByName.size());
before the loop. This ensures that the later push_back() operations never
actually need to reallocate. Unfortunately the check for "is there still spare
capacity" inside push_back() is not optimized away by current compilers.

Version 2 of the above code starts by creating a view on the 'widgetsByName'
range. This view transforms each element (a OSDWidget*) into the corresponding
name of that widget. Then this (lazy) view is materialized into a vector.

The to_vector() function can upfront 'see' how much space is needed for the
vector and thus only needs 1 memory allocation (technically this is possible
because a view::transform() of a random-access range is also a random-access
range). Additionally the copy loop in to_vector() does not need a check for
spare-capacity.

So version 2 is simpler than version 1 and it's also faster, even if we add
reserve(). And to be honest, looking at the existing code, we rarely write that
extra reverse() line. Now it's no longer necessary.