[13/15] Use view::transform() with addListElements()
Similar to the previous patch. Instead of adding elements to a Tcl list
one-by-one, we add a whole range at once. That range is a transformed
view of another range.

In principle there can be performance benefits for using the range approach,
though the current implementation does not take advantage of this yet. (I added
it on my TODO list).