Implement random-access iterators for xrange()
Before this patch the xrange() utility only implemented forward-iterators. Now
that's extended to random-access iterators.

The main motivation for this the following example (from src/GlobalSettings.cc)
    deadzoneSettings = to_vector(
        view::transform(xrange(SDL_NumJoysticks()), [&](auto i) {
            return std::make_unique<IntegerSetting>(
                ...
                strCat("joystick", i + 1, "_deadzone"),
                ...);
            }));

When xrange() only offers forward-iterators, to_vector() has no efficient way
to calculate the number of elements in the range returned by xrange(). And thus
it cannot up-front allocate an internal memory block for the vector with the
correct size. This is possible with random-access iterators.