Followup on ea7d554a4f1 (and a65e265f304e)
a65e265f304e added a workaround for a 'bug'(*) in libstdc++. That bug got
'fixed' in the version that ships with gcc-4.7.2

ea7d554a4f1 removed that workaround because pre-gcc-4.7.2 compilers are no
longer supported.

However Grauw now sees this crash:
  https://github.com/openMSX/openMSX/issues/1136

It seems that libc++ (as opposed to libstdc++) is implementing ~unique_ptr() in
the same 'buggy'(*) way as old libstdc++ versions.

(*) When re-reading the gcc bug report
  http://gcc.gnu.org/bugzilla/show_bug.cgi?id=54351
it's not entirely clear whether this is actually a bug or not.

Therefor I fixed it again in our code, but in a different way:

Problem was that Reactor::getDisplay() was still being called while the Display
is already being destroyed. Old workaround ensured that getDisplay() still
returned a valid pointer (to the object that's being destructed). New code
doesn't call getDisplay() anymore. See a65e265f304e and ea7d554a4f1 for more
details.