[1/2] Fix compile error in view::transform (with libc++)
Grauw reported a compile error on osx after revision cb8c7e4fdc5. I could
reproduce the problem when using clang with -stdlib=libc++ (instead of
libstdc++).

The first (minor, easy to fix) problem was that the reported result type of a
TransformIterator was wrong. It was reported as a reference (T&) but because it
is the result of a calculation (a temporary) it must be a value (T) instead of
a reference (T&). The actual implementation was correct, only the (deprecated?)
reporting mechanism (the 'reference' typedef) was wrong. Possibly libstdc++
does not use this mechanism, while libc++ does?

Unfortunately this only fixed part of the compile errors. The fix for the other
errors is split over the next 3 commits.