Added optional<T> utility
This implements the c++17 std::optional<T> class:
  https://en.cppreference.com/w/cpp/utility/optional
Remove this once we switch to c++17.

This will be used in the next patch to implement semiregular_t<T>. Though it's
a generic utility, we may use it more in the future, possibly to reimplemet
some existing code in a cleaner way.