Added semiregular_t<T> utility
Modeled after the utility with the same name from the "ranges v3" library.
See comments in code for more details.

We need this utility to fix a compile error in view::transform() with
clang/libc++ (this is the default compiler / standard library on osx). More
info on this in the next patch.