[2/2] Fix compile error in view::transform (with libc++)
Grauw reported a compile error on osx after revision cb8c7e4fdc5. I could
reproduce the problem when using clang with -stdlib=libc++ (instead of
libstdc++).

The compile error triggers on a code snippet like this (internally in the
implementation of std::vector in libc++):
    template<typename _ForwardIterator>
    auto insert(..., _ForwardIterator __first, ...) {
        ...
        _ForwardIterator __m = ...;
        ...
        __m = __first; // <--- ERROR
        ...
    }
Here '__m' and '__first' are two objects of the same type (_ForwardIterator),
and we try to assign one to the other. This type is supposed to model the
ForwardIterator concept. And assignment is a valid operation for that concept.

The error message is something like:
    .... error: object of type
    'view::detail::TransformIterator<std::__1::__wrap_iter<const string_view*>,
    (lambda at src/CommandLineParser.cc:95:3)>' cannot be assigned because
    its copy assignment operator is implicitly deleted
   ...

So the concrete type of _ForwardIterator is a view::detail::TransformIterator,
instantiated to transform string_view objects with some lambda function.
Internally a TransformIterator contains another iterator to the to be
transformed range (a string_view-pointer in this case) and the transformation
function (the lambda function in this case). When you assign one
TransformIterator object to another, both sub-parts gets assigned.

The problem is that c++ lambda objects cannot be re-assigned. See here for details:
  https://stackoverflow.com/questions/18755787/redefining-lambdas-not-allowed-in-c11-why

The problem doesn't trigger in libstdc++ because apparently (by luck, or by
design?) that library doesn't re-assign iterators.

To fix this I looked at how other range libraries deal with this problem. The
"ranges v3" library fixes it as follows. Instead of directly storing the
operation (the lambda) in the transform-iterator, they first wrap that operator
in a 'semiregular_t<T>'. That wrapped type is assignable. See previous patch
for more details on semiregular_t. I copied this fix in openMSX.

The cost of this fix is that the operation (the lambda) is now stored in a
optional<T> (an implementation detail of semiregular_t). In practice this means
that the size of a TransformIterator has grown by the size of a boolean
(+padding).