Change TclObject::setXXX() methods to operator=()
Similar cleanup as in the previous two patches, but now for the various
setXXX() methods. Also used the more conventional operator=() for this
functionality.

This also makes it easier to use. For example:
    result.setDouble(myValue / 100.0);
    ...
    result.setInt(myValue);
    ...
    result.setBoolean(myValue < 10);
now simply becomes
    result = myValue / 100.0;
    ...
    result = myValue;
    ...
    result = myValue < 10;
And there no risk anymore of accidentally combining the wrong setXXX() method
with the wrong type.