Renamed signalEvent() to signalMSXEvent()
We had two virtual methods in two different classes:
    int     EventListener::signalEvent(shared_ptr<const Event> event);
    void MSXEventListener::signalEvent(shared_ptr<const Event> event, EmuTime time);
There's one class in our code base (Video9000) that (indirectly) derives from
both EventListener and MSXEventListener. According to c++ that's perfectly
valid: both virtual methods have a different signature, so c++ can distinguish
them.

However clang warns on this. Before the 'override' c++ keyword was introduced
this was a useful warning (because sometimes the different signature is a
mistake). Gcc does not produce this warning (I didn't not check that gcc
silences the warning when it sees that 'override' is used, or that it doesn't
know about this warning at all).

It's easy to work around this. I renamed the 2nd method to 'signalMSXEvent'.
And that's even a slightly more descriptive name anyway.